// src/common/enums/weight-instruction.enum.ts
export enum WeightInstruction {
  BARBELL_SUM = 'barbell_sum',
  MACHINE_DISPLAY = 'machine_display',
  MACHINE_SUM_SIDES = 'machine_sum_sides',
  BODYWEIGHT = 'bodyweight',
  NO_WEIGHT = 'no_weight',
  CABLE_DISPLAY = 'cable_display',
  CABLE_SUM_SIDES = 'cable_sum_sides',
  EACH_SIDE_WEIGHT = 'each_side_weight',
}